# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Foodgram'
  version: ''
paths:
...
  /api/recipes/{id}/get-link/:
    get:
      operationId: Получить короткую ссылку на рецепт
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор рецепта."
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeGetShortLink'
          description: 'Ссылка на рецепт'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
...
  /api/ingredients/:
    get:
      operationId: Список ингредиентов
      description: 'Список ингредиентов с возможностью поиска по имени.'
      parameters:
        - name: name
          required: false
          in: query
          description: Поиск по частичному вхождению в начале названия ингредиента.
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
          description: ''
      tags:
        - Ингредиенты
...
components:
  schemas:
...
    RecipeGetShortLink:
      type: object
      properties:
        short-link:
          type: string
          description: 'Сокращенная ссылка'
          format: uri
          example: 'https://foodgram.example.org/s/3d0'
...
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 128
          example: 'Капуста'
        measurement_unit:
          type: string
          maxLength: 64
          example: 'кг'
      required:
        - name
        - measurement_unit
...
    NotFound:
      description: Объект не найден
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
  responses:
...
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token